Moved current VB state to app.tsx
